🚚 rename :: 컴포넌트 폴더 내 UI 담당 컴포넌트 common으로 분류
