+++ src/refactoring/components/common-ui/layout/CardTemplate.tsx
