+++ src/refactoring/components/common-ui/layout/SectionTemplate.tsx
