+++ src/refactoring/components/user/CartDetail.tsx
@@ -1,7 +1,7 @@
 import { FC } from 'react';
 import { getMaxApplicableDiscount } from '../../hooks/utils/cartUtils';
 import { CartItem } from '../../../types';
-import { Button } from '../atoms';
+import { Button } from '../common-ui/atoms';
 
 type CardDetailProps = {
   item: CartItem;
