+++ src/refactoring/components/user/CouponDropdown.tsx
@@ -1,7 +1,7 @@
 import { FC } from 'react';
 import { Coupon } from '../../../types';
 import { formatKrPrice, discountChangeFormat } from '../../hooks/utils';
-import { Select, SelectOption } from '../atoms/Select';
+import { Select, SelectOption } from '../common-ui/atoms/Select';
 
 type CouponDropdownProps = {
   coupons: Coupon[];
