+++ src/refactoring/components/user/ProductContent.tsx
@@ -1,7 +1,7 @@
 import { FC } from 'react';
 import { getMaxDiscount } from '../../hooks/utils';
 import { Product } from '../../../types';
-import { Button } from '../atoms';
+import { Button } from '../common-ui/atoms';
 
 type ProductContentProps = {
   product: Product;
